Misc tweaks
no functional change

+++ PKHeX.Core/MysteryGifts/WC6.cs
@@ -19,7 +19,7 @@ public sealed class WC6(Memory<byte> raw) : DataMysteryGift(raw), IRibbonSetEven
     public override bool FatefulEncounter => !IsLinkGift; // Link gifts do not set fateful encounter
     public bool IsLinkGift => Location == Locations.LinkGift6;
 
-    public int RestrictLanguage { get; set; } // None
+    public byte RestrictLanguage { get; set; } // None
     public byte RestrictVersion { get; set; } // Permit All
 
     public bool CanBeReceivedByVersion(GameVersion version)
